refactor: split extension commands into modules
